added the description for StorageClass dropdown
