+++ packages/odf/utils/odf.ts
@@ -1,6 +1,8 @@
 import { ODF_OPERATOR } from '@odf/shared/constants';
 import { ClusterServiceVersionKind } from '@odf/shared/types';
 import { K8sResourceKind } from '@odf/shared/types';
+import { StorageClassResourceKind } from '@odf/shared/types';
+import { isDefaultClass } from '@odf/shared/utils';
 import * as _ from 'lodash';
 import { ODF_VENDOR_ANNOTATION } from '../constants';
 
@@ -19,6 +21,19 @@ export const getExternalSubSystemName = (
     0,
     230
   );
+
+export const getStorageClassDescription = (
+  sc: StorageClassResourceKind
+): string => {
+  const descriptionList = [sc.provisioner, sc.parameters?.type];
+  return descriptionList.reduce(
+    (description, dl) => {
+      if (dl) return description ? `${description} | ${dl}` : dl;
+      return description;
+    },
+    isDefaultClass(sc) ? '(default)' : ''
+  );
+};
 
 export const getOperatorVersion = (operator: K8sResourceKind): string =>
   operator?.status?.installedCSV;
